+++ rush-plugins/rush-archive-project-plugin/src/commands/test/archive.test.ts
@@ -1,7 +1,7 @@
 import * as path from "path";
 import { Executable } from "@rushstack/node-core-library";
 import { archive } from "../archive";
-import { graveyardRelativeFolder } from "../../logic/graveyard";
+import { defaultGraveyardRelativeFolder } from "../../logic/graveyard";
 
 const folderSet = new Set<string>();
 
@@ -29,7 +29,7 @@ describe("archive", () => {
     );
     const fixtureMonorepoGraveyardPath = path.join(
       fixtureMonorepoPath,
-      graveyardRelativeFolder
+      defaultGraveyardRelativeFolder
     );
     folderSet.add(fixtureMonorepoPath);
     folderSet.add(fixtureMonorepoGraveyardPath);
@@ -45,6 +45,7 @@ describe("archive", () => {
       await expect(
         archive({
           packageName,
+          gitCheckpoint: false
         })
       ).rejects.toThrow(
         `Could not find project with package name ${packageName}`
@@ -57,6 +58,7 @@ describe("archive", () => {
         await expect(
           archive({
             packageName,
+            gitCheckpoint: false
           })
         ).resolves.toBeUndefined();
       });
@@ -71,7 +73,7 @@ describe("archive", () => {
     );
     const fixtureMonorepoGraveyardPath = path.join(
       fixtureMonorepoPath,
-      graveyardRelativeFolder
+      defaultGraveyardRelativeFolder
     );
     folderSet.add(fixtureMonorepoPath);
     folderSet.add(fixtureMonorepoGraveyardPath);
@@ -87,6 +89,7 @@ describe("archive", () => {
       await expect(
         archive({
           packageName,
+          gitCheckpoint: false
         })
       ).resolves.toBeUndefined();
     });
@@ -100,7 +103,7 @@ describe("archive", () => {
     );
     const fixtureMonorepoGraveyardPath = path.join(
       fixtureMonorepoPath,
-      graveyardRelativeFolder
+      defaultGraveyardRelativeFolder
     );
     folderSet.add(fixtureMonorepoPath);
     folderSet.add(fixtureMonorepoGraveyardPath);
@@ -116,6 +119,7 @@ describe("archive", () => {
       await expect(
         archive({
           packageName,
+          gitCheckpoint: false
         })
       ).rejects.toThrowErrorMatchingSnapshot();
     });