+++ rush-plugins/rush-archive-project-plugin/src/logic/git.ts
@@ -52,3 +52,27 @@ export const gitCheckIgnored = (cwd: string, filePath: string): string => {
   }
   return result;
 };
+
+export const getCheckpointBranch = (cwd: string, branchName: string): string => {
+  const gitPath: string = getGitPathOrThrow();
+  const currDate: string = new Date().toISOString().substring(0, 10);
+  const branchNameToCreate: string = `${branchName}-checkpoint-${currDate}`;
+  const process: SpawnSyncReturns<string> = Executable.spawnSync(gitPath, ["branch", branchNameToCreate], {
+    currentWorkingDirectory: cwd,
+  });
+  if (process.status === 128) {
+    throw new Error(`The passed branch name is invalid: ${branchNameToCreate}`)
+  }
+
+  return branchNameToCreate;
+}
+
+export const pushGitBranch = (cwd: string, branchName: string): void => {
+  const gitPath: string = getGitPathOrThrow();
+  const process: SpawnSyncReturns<string> = Executable.spawnSync(gitPath, ["push", "origin", `${branchName}:${branchName}`], {
+    currentWorkingDirectory: cwd,
+  });
+  if (process.status !== 0) {
+    console.error('Could not push branch to origin');
+  }
+}