+++ rush-plugins/rush-archive-project-plugin/src/logic/graveyard.ts
@@ -1,9 +1,12 @@
-import * as path from "path";
-import { PackageName, IParsedPackageName } from "@rushstack/node-core-library";
+import * as path from 'path';
+import { PackageName, IParsedPackageName, JsonFile, FileSystem } from '@rushstack/node-core-library';
+import { RushConfiguration } from '@rushstack/rush-sdk';
+import { loadRushConfiguration } from './rushConfiguration';
 
-export const graveyardRelativeFolder: string = "common/_graveyard";
+export const defaultGraveyardRelativeFolder: string = 'common/_graveyard';
 
 export interface IGraveyardInfo {
+  graveyardRelativeFolder: string;
   tarballFolder: string;
   tarballRelativeFolder: string;
   tarballName: string;
@@ -14,19 +17,39 @@ export interface IGetGraveyardInfoParams {
   packageName: string;
 }
 
-export const getGraveyardInfo = ({
-  monoRoot,
-  packageName,
-}: IGetGraveyardInfoParams): IGraveyardInfo => {
-  const parsedPackageName: IParsedPackageName = PackageName.parse(packageName);
-  const tarballRelativeFolder: string = path.join(
-    graveyardRelativeFolder,
-    parsedPackageName.scope
+export interface IArchiveConfig {
+  graveyardFolder?: string;
+}
+
+export const getGraveyardInfo = ({ monoRoot, packageName }: IGetGraveyardInfoParams): IGraveyardInfo => {
+  const rushConfiguration: RushConfiguration = loadRushConfiguration();
+  // Check for config file
+  const pluginOptionsJsonFilePath: string = path.join(
+    rushConfiguration.rushPluginOptionsFolder,
+    `rush-archive-project-plugin.json`
   );
+
+  let graveyardRelativeFolder: string = defaultGraveyardRelativeFolder;
+  let archiveConfigs: IArchiveConfig | undefined;
+  try {
+    archiveConfigs = JsonFile.load(pluginOptionsJsonFilePath);
+  } catch (e) {
+    if (!FileSystem.isNotExistError(e as Error)) {
+      throw e;
+    }
+  }
+
+  if (archiveConfigs?.graveyardFolder) {
+    graveyardRelativeFolder = archiveConfigs.graveyardFolder;
+  }
+
+  const parsedPackageName: IParsedPackageName = PackageName.parse(packageName);
+  const tarballRelativeFolder: string = path.join(graveyardRelativeFolder, parsedPackageName.scope);
   const tarballFolder: string = path.join(monoRoot, tarballRelativeFolder);
   return {
+    graveyardRelativeFolder,
     tarballFolder,
     tarballRelativeFolder,
-    tarballName: `${parsedPackageName.unscopedName}.tar.gz`,
+    tarballName: `${parsedPackageName.unscopedName}.tar.gz`
   };
 };