+++ rush-plugins/rush-archive-project-plugin/CHANGELOG.md
@@ -1,6 +1,28 @@
 # Change Log - rush-archive-project-plugin
 
-This log was last generated on Wed, 31 Aug 2022 04:09:13 GMT and should not be manually modified.
+This log was last generated on Wed, 18 Jan 2023 06:59:28 GMT and should not be manually modified.
+
+## 1.1.7
+Wed, 18 Jan 2023 06:59:28 GMT
+
+### Patches
+
+- Add support for graveyard folder in archive plugin config file
+
+## 1.1.6
+Mon, 16 Jan 2023 05:03:53 GMT
+
+### Patches
+
+- Add ability to push git branch and generate md file for archived projects"
+
+## 1.1.5
+Sat, 07 Jan 2023 01:07:03 GMT
+
+### Patches
+
+- Updated to include a git checkpoint branch when archiving a project
+- Update README for archive plugin and add unarchiving support
 
 ## 1.1.4
 Wed, 31 Aug 2022 04:09:13 GMT